feat: fault resilience (#36)
By using multiworld API, fault resilience is enabled. To demonstrate it,
a diamond topology is created for the resnet152 example.

During the development, a deadlock situation is also indentified.
Originally, rank 0 is a sender while other ranks are receivers.
When receiver calls recv function, it needs to fetch a value from remote
TCPStore to conduct recv function. However, the get method for TCPStore
hangs and eventually times out. To mitigate the issue, the role of rank 0
is changed as receiver; therefore, relevant code and example configs are
also updated accordingly.