feat: tensor meta info via control channel (#50)
Tensor's meta information (shape and dtype) needs to be shared with a tensor
receiver before the receiver can receive the tensor. Currently, the sharing
functionality was implemented as part of data plane (i.e., nccl or gloo).
This is quite rigid and not easy to handle a case where tensor's shape and
dtype can change dynamically. For example, in LLM, the length of prompts can
be difference across prompts, which leads to different sizes of tensor.

In addition, the meta info for multiple tensors need to be shared too.
For example, if the lengths of prompts in a batch for LLM are different,
an LLM needs two inputs as tensor: input_ids and attention_mask. The meta
information is represented as a dictionary of tensors. Based on the info,
when tensors are actually received in the data plane, they are reconstructed
as dictionary and passed to the model.

Because of the above change, the way we shard the model has to be modified too.
Previously, the output is specified as a tensor during the sharding. Now, the
output is represented as a dictionary where key is a string and value is a
tensor.