feat: introduced agent IP address (#67)
Added agent IP address in the context of a job. When registering into the controller, Agent passes it's IP address and that is stored in the Agent's context.
Whenever a start job command is being received, the controller updates the workers addr field with the Agent's IP address on which these workers will be deployed.