feat: added condition for running state transition (#80)
Added method to listen for worker status updates. Whenever a new status update message is received from the worker, job state class handles that message.
For transition to running state, we check if all the workers from a job are in running state. After this check, we transition the job state from STARTING to RUNNING.