feat: plumbing between autoscaler and other components (#168)
In order to do autoscaling, performance metrics need to be stored in
the job context first. Then, autoscaler needs to be informed that
the performance metrics for a server are updated so that it can decide
whether it needs to do autoscaling or not. The basic plumbing to
integrate autoscaler and other components are implemented in this
commit.