fix: rollback resources on error (#181)
Whenever an error while processing a config happens, there might be a case when the devices that were already assign to workers have an incorrect state being set as used but with no job running. We keep records of these devices and to the cleanup whenever an exception is raised. Static policy was updated and wrapped split method with a try / except block in order to revert the devices set as used whenever there's an error.