feat: revert suspended worlds (#261)
Whenever job is about to transition back to running, a new check loop command is being sent to agents with an empty set of failed worker ids. Based on this, the router of each worker will move the suspended worlds back into active data structure so they can be used by the forwarding policy. Updated cond_running and cond_updated methods to handle the command and moved the check loop command send logic into a separate method so it can be called from multiple places.