fix: peer connection error (#273)
When a worker fails, other worker that is connected to the failed one will receive b"" (EOF). For that reason, we added a check for this scenario and when that happens, we also close the connection. When the worker is recovered, the connection is re-created. Previously, we only tried to convert the data to an integer, therefore the ValueError was raised.