feat: worker abort configure (#287)
Worker needs to properly handle back to back configs. For that reason, we implemented an abort mechanism for worker. Whenever there's a configuration task in progress, if a new config is being received, we conditionally cancel current config task. The decision to cancel the current task is made based on computing a diff between new and current spec. Created ConfigRunner class to encapsulate this logic.