feat: limit deploy recovery attempts (#319)
We need to limit back to back failures while trying to recover some workers. For that, we added logic to cancel and start a new recovery task and always using latest set of failed worker ids. Workers will take care of back to back configs by cancelling or continuing configurations. Also, when the recovery deploy limit is reached, we remove the affected pipelines based on failed worker ids and do another update. Logic in config runner was updated to include removed worlds in the cancellation checking process. Added a todo item to remove pipelines from planner as well. We need to keep the planner data structure up to date.